Add files via upload
### Notebook Overview
In this notebook, I will delve into the realm of transfer learning by harnessing the power of pre-trained models from TensorFlow Hub. To expedite experimentation, I've chosen to work with only 10% of the dataset, offering a glimpse into the world of rapid prototyping. This journey into transfer learning will encompass the following key areas:

### 1. Transfer Learning with TensorFlow
I'll explore the fundamental concept of transfer learning and its practical application using TensorFlow, shedding light on how pre-trained models can be leveraged for various tasks.

### 2. Leveraging a Reduced Dataset
To streamline our experimentation, I've opted to work with a mere 10% of the training samples. This approach accelerates the model development process while providing valuable insights into its performance.

### 3. Building a Transfer Learning Feature Extraction Model
Using TensorFlow Hub, I'll demonstrate how to construct a feature extraction model for transfer learning. We'll focus on two specific pre-trained models:

   - EfficientB0
   - ResNet50

### 4. Tracking Progress with TensorBoard
To keep a watchful eye on our modeling experiments and their outcomes, I'll introduce the use of TensorBoard. It's a vital tool for visualizing and evaluating the performance of our models. This notebook's primary objective is to showcase the distinctions between a conventional CNN and a pre-trained one, culminating in a comparative analysis of the results. Spoiler alert: EfficientB0 emerges as the winner in terms of accuracy and computational efficiency.